Fix a few warnings; remove unused code.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@5317 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ protannot/src/org/bioviz/protannot/GenomeView.java
@@ -144,7 +144,7 @@ final class GenomeView extends JPanel implements MouseListener{
     private static final int divider_size = 8;
     private static final int table_height = 150;
     private static final int seqmap_pixel_height = 500;
-	private static final int zoomRatio = 30;
+	private static final double zoomRatio = 30.0;
     
     /**
      * Removes currently loaded data by clearing maps.
